Toutes mes condoléhence morgan

+++ lib/features/Products/app/helpers/product_reference_manager_helper.rb
@@ -5,7 +5,7 @@ module ProductReferenceManagerHelper
     categories = ProductReferenceCategory.find_all_by_product_reference_category_id
     list = []
     list << "<div class='hierarchic'>"
-    list << "<ul class='parent'>"
+    list << "<ul class=\"parent\">"
     list = get_children(categories,list)
     list << "</ul>"
     list << "</div>"
@@ -16,21 +16,24 @@ module ProductReferenceManagerHelper
   def get_children(categories,list)
     categories.each do |category|
       delete_button = show_delete_button(category)
-      list << "<li class='category'>#{category.name} &nbsp; (#{category.product_references_count})&nbsp; <span class='action'>"+
+      list << "<li class=\"category\">#{category.name} &nbsp; (#{category.product_references_count})&nbsp; <span class=\"action\">"+
         link_to( 'Ajouter une cat&eacute;gorie' , new_product_reference_category_path(:id => category.id) )+" &brvbar; "+
         link_to( 'Ajouter une r&eacute;f&eacute;rence' , new_product_reference_path(:id => category.id) )+" &brvbar; "+
         link_to( 'Modifier', edit_product_reference_category_path(category) )+"&nbsp; #{delete_button}</span></li>"
-      list << "<ul>"
-      if category.children.size > 0
-        get_children(category.children,list)
-      end
-      if category.product_references.size > 0
-        category.product_references.each do |reference|
-          list << "<li class='reference'>#{reference.name} &nbsp; (#{reference.products_count})&nbsp; <span class='action'>"+
-            link_to( 'Modifier', edit_product_reference_path(reference) )+" &brvbar; "+link_to("Supprimer", reference, { :method => :delete})+"</span></li>"          
+
+      if category.children.size > 0 or category.product_references.size > 0
+        list << "<li><ul>"
+        if category.children.size > 0
+          get_children(category.children,list)
         end
+        if category.product_references.size > 0
+          category.product_references.each do |reference|
+            list << "<li class=\"reference\">#{reference.name} &nbsp; (#{reference.products_count})&nbsp; <span class=\"action\">"+
+              link_to( 'Modifier', edit_product_reference_path(reference) )+" &brvbar; "+link_to("Supprimer", reference, { :method => :delete})+"</span></li>"          
+          end
+        end
+        list << "</ul></li>"
       end
-      list << "</ul>"
     end
     list
   end
